+++ libs/ledger-live-common/src/families/near/api/index.ts
@@ -1,0 +1,13 @@
+export { getOperations } from "./datahub-indexer-sdk";
+export {
+  getAccount,
+  fetchAccountDetails,
+  getProtocolConfig,
+  getGasPrice,
+  getAccessKey,
+  broadcastTransaction,
+  getStakingPositions,
+  getValidators,
+  getCommission,
+} from "./archive-node-sdk";
+export { getStakingDeposits } from "./near-staking-positions-sdk";